+++ assistant_webserver/auth.py
@@ -10,3 +10,9 @@ def auth_error(username, password):
         return -2
     except KeyError:
         return -1
+
+
+def get_coords(name):
+    if name == "Brad Thaxton":
+        coords = {'x': -2.16564114295, 'y': -5.48124138424}
+        return coords
