+++ assistant_webserver/templates/main.html
@@ -44,7 +44,7 @@
 <div class="white-div">
 <h1>Hi Friend!</h1>
 <p>Do you need help?<br>I can help you find required person:</p>
-<form id="f1" method="POST" action="/login">
+<form id="f1" method="POST" action="/search">
     <input name="user_name" type="text" placeholder="Login"><br>
     <input name="password" type="password" id="p1" placeholder="Password"><br>
